putting in a guard for git hook installation. it's possible the user did not clone the repo but downloaded a zip of it instead. the git hook would fail in that case, now we guard against it

+++ regis/git_hooks.py
@@ -9,11 +9,15 @@ def __install(hooksPath):
     if not os.path.exists(hooksPath):
         return
     
+    git_path = os.path.join(root_path, ".git")
+    if not os.path.exist(git_path):
+        return
+    
     hooks = os.listdir(hooksPath)
 
     for hook in hooks:
         src = os.path.join(hooksPath, hook)
-        dst = os.path.join(root_path, ".git", "hooks", hook)
+        dst = os.path.join(git_path, "hooks", hook)
         shutil.copy(src, dst)
 
 def run(hooksPath):
